Do not hard-code pipelineSdkVersion in pipeline
Do not hard-code pipelineSdkVersion in order to allow control of the version from -
 * jenkins.yml, or
 * Manage Jenkins / Configure System / Global Pipeline Libraries.
This makes it easier to switch to a custom (forked) version of the library, where the version is e.g. 'patch-1'.
Currently when trying to force a version with 'Allow default version to be overridden' off, there is an error in the 'Init' stage:
"Version override not permitted for library s4sdk-pipeline-library".